fast_api/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
GeoAlchemy2==0.9.4
oauthlib==3.1.1
requests-oauthlib==1.3.0
sentinelhub==3.4.1
shapely
geojson
shapely
celery
tortoise-orm[asyncpg]
python-decouple
shapely
geopandas